chore(version): move version package to pkg
